move signature to classname
